backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pillow==11.0.0
numpy==2.2.1
pandas==2.2.3
openpyxl==3.1.5
pytest==8.3.4
httpx==0.28.1
